Changed how I handle upvotes and downvotes
On a patch request I check the status of the request and find out if
the user making the request is within upvote users/ downvote users. I
change the votetype and send this back as a reference for the front
end to choose how to visually show the changes.
example:

If a user requests an upvote and they already exist in downvote, I
pull them out of the downvote array and push them into upvote array and
send votetype getridofdownvoteandaddupvote